+++ processor/src/test/fixtures/input/com/example/component/UnnecessaryDefaultSkipIfDisposedComponentModel.java
@@ -1,0 +1,9 @@
+package com.example.component;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.Feature;
+
+@ArezComponent( defaultSkipIfDisposed = Feature.ENABLE, allowEmpty = true )
+abstract class UnnecessaryDefaultSkipIfDisposedComponentModel
+{
+}